+++ src/scss/components.scss
@@ -1,6 +1,6 @@
 // Components
 @import '../components/Element/style';
-@import '../components/InputField/style';
+@import '../components/InputElement/style';
 @import '../components/TextAreaInput/style';
 @import '../components/NumericInput/style';
 @import '../components/SliderInput/style';
